finished refactoring node.parent to point to a node, NodeImplTest is finished. more tests on the way.

+++ core/src/main/java/org/devilry/core/session/TreeManagerImpl.java
@@ -19,7 +19,7 @@ public class TreeManagerImpl implements TreeManagerRemote {
 		Node node = new Node();
 		node.setName(name.toLowerCase());
 		node.setDisplayName(displayName);
-		node.setParentId(parentId);
+		node.setParent(getNode(parentId));
 
 		em.persist(node);
 		em.flush();
@@ -37,7 +37,7 @@ public class TreeManagerImpl implements TreeManagerRemote {
 		node.setName(name.toLowerCase());
 		node.setDisplayName(displayName);
 		node.setCourseCode(courseCode);
-		node.setParentId(parentId);
+		node.setParent(getNode(parentId));
 
 		em.persist(node);
 		em.flush();
@@ -56,7 +56,7 @@ public class TreeManagerImpl implements TreeManagerRemote {
 		node.setDisplayName(displayName);
 		node.setStartDate(start);
 		node.setEndDate(end);
-		node.setParentId(parentId);
+		node.setParent(getNode(parentId));
 
 		em.persist(node);
 		em.flush();
@@ -69,7 +69,7 @@ public class TreeManagerImpl implements TreeManagerRemote {
 		AssignmentNode node = new AssignmentNode();
 		node.setName(name.toLowerCase());
 		node.setDisplayName(displayName);
-		node.setParentId(parentId);
+		node.setParent(getNode(parentId));
 
 		em.persist(node);
 		em.flush();
@@ -80,7 +80,11 @@ public class TreeManagerImpl implements TreeManagerRemote {
 	@TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 	public void delNode(long nodeId) {
 		Node node = getNode(nodeId);
-		em.remove(node);
+
+		if(node != null) {
+			em.remove(node);
+			em.flush();
+		}
 	}
 
 	public long getNodeIdFromPath(String path) {
@@ -119,11 +123,11 @@ public class TreeManagerImpl implements TreeManagerRemote {
 		Query q;
 		
 		if(parentId != -1) {
-			q = em.createQuery("SELECT n FROM Node n WHERE n.name=:name AND n.parentId=:parentId");
+			q = em.createQuery("SELECT n FROM Node n WHERE n.name=:name AND n.parent IS NOT NULL AND n.parent.id=:parentId");
 			q.setParameter("name", name);
 			q.setParameter("parentId", parentId);
 		} else {
-			q = em.createQuery("SELECT n FROM Node n WHERE n.name=:name AND n.parentId=-1");
+			q = em.createQuery("SELECT n FROM Node n WHERE n.name=:name AND n.parent IS NULL");
 			q.setParameter("name", name);
 		}
 
@@ -140,37 +144,25 @@ public class TreeManagerImpl implements TreeManagerRemote {
 
 	private long getNodeId(String name, String parent) {
 		Query q;
-
+		
 		if(parent != null) {
-			q = em.createQuery("SELECT n FROM Node n WHERE n.name=:name AND n.parentId<>-1");
+			q = em.createQuery("SELECT n FROM Node n WHERE n.name=:name AND n.parent IS NOT NULL AND n.parent.name=:parent");
 			q.setParameter("name", name);
-
-			for(Node n : (java.util.List<Node>) q.getResultList()) {
-				Query pq = em.createQuery("SELECT n.name FROM Node n WHERE n.id=:parentId");
-				pq.setParameter("parentId", n.getParentId());
-
-				try {
-					String pname = (String) pq.getSingleResult();
-					if(parent.equals(pname))
-						return n.getId();
-
-				} catch(NoResultException e) {
-					return -1;
-				}
-			}
-
-			return -1;
+			q.setParameter("parent", parent);
 		} else {
-			q = em.createQuery("SELECT n FROM Node n WHERE n.name=:name AND n.parentId=-1");
+			q = em.createQuery("SELECT n FROM Node n WHERE n.name=:name AND n.parent IS NULL");
 			q.setParameter("name", name);
-
-			try {
-				Node node = (Node) q.getSingleResult();
-				return node.getId();
-			} catch(NoResultException e) {
-				return -1;
-			}
 		}
+
+		Node node;
+
+		try {
+			node = (Node) q.getSingleResult();
+		} catch(NoResultException e) {
+			node = null;
+		}
+
+		return node==null?-1:node.getId();
 	}
 }
